+++ cmd/traffic/cmd/manager/internal/cluster/info.go
@@ -13,7 +13,6 @@ import (
 	"github.com/datawire/ambassador/pkg/kates"
 	"github.com/datawire/dlib/derror"
 	"github.com/datawire/dlib/dlog"
-	"github.com/telepresenceio/telepresence/rpc/v2/manager"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/manager"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/iputil"
@@ -31,12 +30,14 @@ type info struct {
 	Pods    []kates.Pod
 	Nodes   []kates.Node
 	waiter  sync.Cond
+
+	// podCIDRMap keeps track of the current set of pod CIDRs
+	podCIDRMap map[iputil.IPKey]int
 }
 
 func NewInfo(ctx context.Context) Info {
 	oi := info{}
 	oi.waiter.L = &oi.accLock
-	oi.PodSubnets = make([]*rpc.IPNet, 0, 2)
 
 	client := managerutil.GetKatesClient(ctx)
 	if client == nil {
@@ -115,17 +116,18 @@ func NewInfo(ctx context.Context) Info {
 		oi.ServiceSubnet = &rpc.IPNet{Ip: net.IP(oi.KubeDnsIp).Mask(mask), Mask: int32(ones)}
 	}
 
-	oi.PodSubnets = oi.getPodCIDRsFromNodes(ctx)
-	if len(oi.PodSubnets) == 0 {
+	podCIDRs := oi.getPodCIDRsFromNodes(ctx)
+	if len(podCIDRs) == 0 {
 		// Some clusters (e.g. Amazon EKS) doesn't assign podCIDRs to nodes
 		// by default so we compute those CIDRs by looking at all pods instead.
 		dlog.Infof(ctx, "Deriving subnets from IPs of pods")
-		oi.PodSubnets = oi.getPodCIDRsFromPods(ctx)
+		podCIDRs = oi.getPodCIDRsFromPods(ctx)
 		oi.watchSubnets(ctx, "Pods", "pod", oi.podCIDRsFromPods)
 	} else {
 		dlog.Infof(ctx, "Deriving subnets from podCIRs of nodes")
 		oi.watchSubnets(ctx, "Nodes", "node", oi.podCIDRsFromNodes)
 	}
+	oi.updateSubnets(podCIDRs)
 	return &oi
 }
 
@@ -165,7 +167,7 @@ func (oi *info) clusterInfo() *rpc.ClusterInfo {
 	return ci
 }
 
-func (oi *info) watchSubnets(ctx context.Context, name, kind string, retriever func(context.Context) []*manager.IPNet) {
+func (oi *info) watchSubnets(ctx context.Context, name, kind string, retriever func(context.Context) []*net.IPNet) {
 	acc := managerutil.GetKatesClient(ctx).Watch(ctx,
 		kates.Query{
 			Name: name,
@@ -178,13 +180,36 @@ func (oi *info) watchSubnets(ctx context.Context, name, kind string, retriever f
 			case <-ctx.Done():
 				return
 			case <-acc.Changed():
-				if oi.accUpdate(ctx, acc) {
-					oi.PodSubnets = retriever(ctx)
+				if oi.accUpdate(ctx, acc) && oi.updateSubnets(retriever(ctx)) {
 					oi.waiter.Broadcast()
 				}
 			}
 		}
 	}()
+}
+
+func (oi *info) updateSubnets(podCIDRs []*net.IPNet) bool {
+	changed := true
+	if len(podCIDRs) == len(oi.podCIDRMap) {
+		changed = false // assume equal
+
+		// Check if all IPs are found and that their masks are equal
+		for _, cidr := range podCIDRs {
+			if ones, ok := oi.podCIDRMap[iputil.IPKey(cidr.IP)]; ok {
+				newOnes, _ := cidr.Mask.Size()
+				if newOnes == ones {
+					continue
+				}
+			}
+			changed = true
+			break
+		}
+	}
+	if changed {
+		oi.podCIDRMap = makeCIDRMap(podCIDRs)
+		oi.PodSubnets = toRPCSubnets(podCIDRs)
+	}
+	return changed
 }
 
 // accUpdate updates the relevant field in the info and recovers any panics that
@@ -200,7 +225,7 @@ func (oi *info) accUpdate(ctx context.Context, a *kates.Accumulator) bool {
 	return a.Update(oi)
 }
 
-func (oi *info) getPodCIDRsFromNodes(ctx context.Context) []*rpc.IPNet {
+func (oi *info) getPodCIDRsFromNodes(ctx context.Context) []*net.IPNet {
 	if err := managerutil.GetKatesClient(ctx).List(ctx, kates.Query{Kind: "Node"}, &oi.Nodes); err != nil {
 		dlog.Errorf(ctx, "failed to get nodes: %v", err)
 		return nil
@@ -208,8 +233,8 @@ func (oi *info) getPodCIDRsFromNodes(ctx context.Context) []*rpc.IPNet {
 	return oi.podCIDRsFromNodes(ctx)
 }
 
-func (oi *info) podCIDRsFromNodes(ctx context.Context) []*rpc.IPNet {
-	subnets := make([]*rpc.IPNet, 0, len(oi.Nodes))
+func (oi *info) podCIDRsFromNodes(ctx context.Context) []*net.IPNet {
+	subnets := make([]*net.IPNet, 0, len(oi.Nodes))
 	for i := range oi.Nodes {
 		node := &oi.Nodes[i]
 		spec := node.Spec
@@ -223,14 +248,13 @@ func (oi *info) podCIDRsFromNodes(ctx context.Context) []*rpc.IPNet {
 				dlog.Errorf(ctx, "unable to parse podCIDR %q in node %s", cs, node.Name)
 				continue
 			}
-			dlog.Infof(ctx, "Using podCIDR %s from node %s", cidr, node.Name)
-			subnets = append(subnets, iputil.IPNetToRPC(cidr))
+			subnets = append(subnets, cidr)
 		}
 	}
 	return subnets
 }
 
-func (oi *info) getPodCIDRsFromPods(ctx context.Context) []*rpc.IPNet {
+func (oi *info) getPodCIDRsFromPods(ctx context.Context) []*net.IPNet {
 	if err := managerutil.GetKatesClient(ctx).List(ctx, kates.Query{Kind: "Pod"}, &oi.Pods); err != nil {
 		dlog.Errorf(ctx, "failed to get pods: %v", err)
 		return nil
@@ -238,7 +262,7 @@ func (oi *info) getPodCIDRsFromPods(ctx context.Context) []*rpc.IPNet {
 	return oi.podCIDRsFromPods(ctx)
 }
 
-func (oi *info) podCIDRsFromPods(ctx context.Context) []*rpc.IPNet {
+func (oi *info) podCIDRsFromPods(ctx context.Context) []*net.IPNet {
 	ips := make(iputil.IPs, 0, len(oi.Pods))
 	for i := range oi.Pods {
 		pod := &oi.Pods[i]
@@ -256,7 +280,18 @@ func (oi *info) podCIDRsFromPods(ctx context.Context) []*rpc.IPNet {
 			ips = append(ips, ip)
 		}
 	}
-	cidrs := subnet.CoveringCIDRs(ips)
+	return subnet.CoveringCIDRs(ips)
+}
+
+func makeCIDRMap(cidrs []*net.IPNet) map[iputil.IPKey]int {
+	m := make(map[iputil.IPKey]int, len(cidrs))
+	for _, cidr := range cidrs {
+		m[iputil.IPKey(cidr.IP)], _ = cidr.Mask.Size()
+	}
+	return m
+}
+
+func toRPCSubnets(cidrs []*net.IPNet) []*rpc.IPNet {
 	subnets := make([]*rpc.IPNet, len(cidrs))
 	for i, cidr := range cidrs {
 		subnets[i] = iputil.IPNetToRPC(cidr)